Address Adam's comments
