+++ src/dust/dust.cpp
@@ -248,13 +248,12 @@ Real EstimateTimestepMesh(MeshData<Real> *md) {
 
   // NOTE(@pdmullen): Without FARGO, dt must be additionally limited by the linear
   // advection of the shear background flow (vy0 = -q Omega x)
-  bool do_shear = false;
   Real qshear = 0.0, om0 = 0.0;
-  if (pm->packages.Get("artemis")->Param<bool>("do_rotating_frame")) {
+  const bool do_shear = pm->packages.Get("artemis")->Param<bool>("do_shear");
+  if (do_shear) {
     auto &rframe_pkg = pm->packages.Get("rotating_frame");
     qshear = rframe_pkg->Param<Real>("qshear");
     om0 = rframe_pkg->Param<Real>("omega");
-    do_shear = (qshear * om0 != 0.0);
   }
 
   static auto desc =
