+++ src/radiation/imc/imc.hpp
@@ -15,24 +15,11 @@
 
 // Artemis includes
 #include "artemis.hpp"
-#include "derived/fill_derived.hpp"
-
-// Jaybenne includes
-#include "jaybenne.hpp"
-
-using namespace parthenon::driver::prelude;
 
 namespace IMC {
-//----------------------------------------------------------------------------------------
-//! \fn TaskListStatus IMC::JaybenneIMC
-//! \brief Executes thermal IMC transport (Jaybenne) and syncs updated fields
+
 template <Coordinates GEOM>
-TaskListStatus JaybenneIMC(Mesh *pmesh, const Real time, const Real dt) {
-  auto status = jaybenne::RadiationStep(pmesh, time, dt).Execute();
-  if (status != TaskListStatus::complete) return status;
-  status = ArtemisDerived::SyncFields<GEOM>(pmesh, time, dt).Execute();
-  return status;
-}
+TaskListStatus JaybenneIMC(Mesh *pmesh, const Real time, const Real dt);
 
 } // namespace IMC
